+++ src/main/kotlin/com/dooingle/domain/catch/model/Catch.kt
@@ -11,7 +11,6 @@ class Catch(
     @Column
     var content: String,
 
-//    @ManyToOne(fetch = FetchType.LAZY)
     @OneToOne
     @JoinColumn(name = "dooingle_id")
     val dooingle: Dooingle,
