+++ src/main/kotlin/com/dooingle/domain/dooingle/dto/DooingleResponse.kt
@@ -1,5 +1,7 @@
 package com.dooingle.domain.dooingle.dto
 
+import com.dooingle.domain.catch.dto.CatchResponse
+import com.dooingle.domain.catch.model.Catch
 import com.dooingle.domain.dooingle.model.Dooingle
 import java.time.ZonedDateTime
 
@@ -18,5 +20,20 @@ data class DooingleResponse(
                 createdAt = dooingle.createdAt
             )
         }
+    }
+
+    fun toDooingleAndCatchResponse(catch: Catch?): DooingleAndCatchResponse {
+        return DooingleAndCatchResponse(
+            ownerName = ownerName,
+            dooingleId = dooingleId,
+            content = content,
+            catch = (catch)?.let {
+                if (it.deletedAt == null)
+                    CatchResponse.from(it)
+                else
+                    "삭제된 캐치입니다."
+            },
+            createdAt = createdAt
+        )
     }
 }
