+++ src/main/kotlin/com/dooingle/domain/dooingle/repository/DooingleQueryDslRepository.kt
@@ -1,8 +1,6 @@
 package com.dooingle.domain.dooingle.repository
 
 import com.dooingle.domain.dooingle.dto.DooingleResponse
-import com.dooingle.domain.dooingle.dto.DooingleAndCatchResponse
-import com.dooingle.domain.dooingle.dto.DooingleQueryResponse
 import com.dooingle.domain.user.model.User
 import org.springframework.data.domain.Pageable
 import org.springframework.data.domain.Slice
@@ -11,7 +9,7 @@ interface DooingleQueryDslRepository {
 
     fun getDooinglesBySlice(cursor: Long?, pageable: Pageable): Slice<DooingleResponse>
 
-    fun getPersonalPageBySlice(owner: User, cursor: Long?, pageable: Pageable): Slice<DooingleQueryResponse>
+    fun getPersonalPageBySlice(owner: User, cursor: Long?, pageable: Pageable): Slice<DooingleResponse>
 
     // TODO 팔로우 기능 구현 후 구현 필요
     // fun getDooinglePageableOfFollows(pageable: Pageable): Slice<Dooingle>
