+++ src/main/kotlin/com/dooingle/global/SecurityConfig.kt
@@ -1,0 +1,22 @@
+package com.dooingle.global
+
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.security.config.annotation.web.builders.HttpSecurity
+import org.springframework.security.web.SecurityFilterChain
+
+@Configuration
+class SecurityConfig {
+    @Bean
+    fun filterChain(http: HttpSecurity): SecurityFilterChain {
+        return http
+            .httpBasic { it.disable() }
+            .formLogin { it.disable() }
+            .csrf { it.disable() }
+            .headers { it.frameOptions { frameOptionsConfig -> frameOptionsConfig.sameOrigin() } }
+            .authorizeHttpRequests {
+                it.anyRequest().permitAll()
+            }
+            .build()
+    }
+}
